Drop MacOS support for local dev environment
I was unable to create a way that plays nice with Mac OS and Linux without creating lots of complexity. Because the actual environment is also focussed running on a Linux host I decided to favour this for a local setup. We should reintroduce a setup for MacOS once I need it again.